update line 129
  for (int i = 0; i < SHA256::DIGEST_SIZE; i++)
change to
  for (int i = 0; i < (int) SHA256::DIGEST_SIZE; i++)

+++ sha256.cpp
@@ -126,7 +126,7 @@ std::string sha256(std::string input)
 
     char buf[2*SHA256::DIGEST_SIZE+1];
     buf[2*SHA256::DIGEST_SIZE] = 0;
-    for (int i = 0; i < SHA256::DIGEST_SIZE; i++)
+    for (int i = 0; i < (int) SHA256::DIGEST_SIZE; i++)
         sprintf(buf+i*2, "%02x", digest[i]);
     return std::string(buf);
 }
